Minor bug fix
The global search feature retains some records on the page even if there
is no match for the string entered. The 'There is no data to display!'
message was also retained. The records displayed were the records
returned by the search function for the longest substring of the search
text that returned records during the search.